pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "matter-web-controller"
version = "0.25.0"
description = "Using websocket matter server to query/control matter-enabled lights/sensors via a HTTP GET interface."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "aiohttp",
    "home-assistant-chip-core",
    "home-assistant-chip-clusters",
    "cryptography",    
    "python-matter-server",
    "zeroconf",
    "mcp[cli]",
]

[project.scripts]
matter-srv = "cli.server:main"
matter-mcp = "cli.mcp_server:main"

[tool.isort]
profile = "black"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
